Teh Amount you want to put on your ProPay Account can now be used!

+++ src/main/java/hhu/propra2/javageddon/teils/model/Reservations.java
@@ -5,6 +5,6 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 @JsonIgnoreProperties(ignoreUnknown = true)
 
 public class Reservations {
-    private int amount;
+    private double amount;
     private int id;
 }
